+++ tests/order/pages/__init__.py
@@ -1,1 +1,1 @@
-__author__ = 'ayia'
+__author__ = 'yashaka'