+++ tests/todomvc/custom_selements_and_collections_test.py
@@ -1,4 +1,4 @@
-__author__ = 'ayia'
+__author__ = 'yashaka'
 
 from selenium import webdriver